[ALPS04316962] cmdq: Cast ptr for user space va
Cast va between 32 and 64bit from user space.

MTK-Commit-Id: 16821ed4a26f3c52bd55a64226feecf31ee4db58

CR-Id: ALPS04316962
Feature: [Module]CMDQ Driver

+++ drivers/misc/mediatek/cmdq/v3/cmdq_driver.c
@@ -358,7 +358,8 @@ static long cmdq_driver_create_secure_medadata(struct cmdqCommandStruct *pComman
 	void *meta_buf;
 	s32 ret;
 	void *addr_meta = CMDQ_U32_PTR(pCommand->secData.addrMetadatas);
-	void *isp_bufs[ARRAY_SIZE(pCommand->secData.ispMeta.ispBufs)] = {0};
+	void *isp_bufs[ARRAY_SIZE(
+		pCommand->secData.ispMeta.ispBufs)] = {0};
 	u32 i;
 
 	max_length = CMDQ_IWC_MAX_ADDR_LIST_LENGTH * sizeof(struct cmdqSecAddrMetadataStruct);
@@ -367,7 +368,8 @@ static long cmdq_driver_create_secure_medadata(struct cmdqCommandStruct *pComman
 	/* always clear to prevent free unknown memory */
 	pCommand->secData.addrMetadatas = 0;
 	for (i = 0; i < ARRAY_SIZE(pCommand->secData.ispMeta.ispBufs); i++) {
-		isp_bufs[i] = (void *)pCommand->secData.ispMeta.ispBufs[i].va;
+		isp_bufs[i] = (void *)(unsigned long)
+			pCommand->secData.ispMeta.ispBufs[i].va;
 		pCommand->secData.ispMeta.ispBufs[i].va = 0;
 	}
